+++ e-httpRequest/CHANGELOG.md
@@ -4,6 +4,9 @@ E-HttpRequest
 v2.0.0
 ---
 * Updated to use UnifiedViews helpers 3.X with rdf4j support
+* Support for HTTPS
+* Support for dynamic configuration (only request for raw mode)
+* Content type may be customized in the file mode
 
 v1.0.0
 ---